Fix compilation of Ext2Create under OpenBSD: off_t casted to long long unsigned. It seems that on OpenBSD the off_t variable is an 64-bit integer on Linux it's 32-bit. We can fix this by casting it to 64-bit, and using %llu in the format.
git-svn-id: https://freenos.googlecode.com/svn/trunk@182 824000e8-63cf-11de-a6e1-e33d94ba4d19

+++ srv/filesystem/ext2/Ext2Create.cpp
@@ -84,8 +84,9 @@ Ext2Inode * Ext2Create::createInode(char *inputFile, Size *number)
         exit(EXIT_FAILURE);
     }
     /* Debug out. */
-    printf("%s mode=%x size=%lu userId=%u groupId=%u\r\n",
-            inputFile, st.st_mode, st.st_size, st.st_uid, st.st_gid);
+    printf("%s mode=%x size=%llu userId=%u groupId=%u\r\n",
+            inputFile, st.st_mode, (long long unsigned)st.st_size,
+	    st.st_uid, st.st_gid);
 
     /* Grab a new inode, or use the given number. */
     if (*number)
